WEBAPP-652: Restructure build configs
