--- build-configs/configs/featureFlags.js
@@ -1,13 +1,0 @@
-// @flow
-
-export type FeatureFlagsType = {|
-  pois: boolean,
-  newsStream: boolean
-|}
-
-const featureFlags: FeatureFlagsType = {
-  pois: false,
-  newsStream: true
-}
-
-export default featureFlags
