+++ build-configs/integreat-test-cms/assets/favicon.ico
