+++ build-configs/integreat-test-cms/assets/manifest.json
