+++ build-configs/integreat/assets/apple-touch-icon.png
