+++ build-configs/integreat/assets/favicon-128x128.png
