+++ build-configs/integreat/assets/favicon-16x16.png
