+++ build-configs/integreat/assets/favicon-192x192.png
