+++ build-configs/integreat/assets/favicon-48x48.png
