+++ build-configs/integreat/assets/favicon-512x512.png
