+++ build-configs/integreat/assets/favicon-96x96.png
