+++ build-configs/integreat/assets/favicon.ico
