+++ build-configs/integreat/assets/integreat-app-logo.png
