+++ build-configs/integreat/assets/manifest.json
