+++ build-configs/integreat/theme/colors.js
@@ -1,9 +1,9 @@
 // @flow
 
-import { commonColors, commonDarkColors } from '../common/colors'
-import type { ColorsType } from '../common/colors'
+import { commonColors, commonDarkColors } from '../../common/theme/colors'
+import type { ColorsType } from '../../common/theme/colors'
 
-const themeColor = '#ff0000'
+const themeColor = '#fbda16'
 
 export const darkColors: ColorsType = {
   themeColor,
