+++ build-configs/integreat/theme/index.js
@@ -2,9 +2,9 @@
 
 import integreatColors, { darkColors as integreatDarkColors } from './colors'
 import integreatFonts from './fonts'
-import dimensions from '../common/dimensions'
-import helpers from '../common/helpers'
-import type { ThemeType } from '../ThemeType'
+import dimensions from '../../common/theme/dimensions'
+import helpers from '../../common/theme/helpers'
+import type { ThemeType } from '../../ThemeType'
 
 const theme: ThemeType = {
   colors: integreatColors,
