+++ build-configs/malte/assets/apple-touch-icon.png
