+++ build-configs/malte/assets/favicon-128x128.png
