+++ build-configs/malte/assets/favicon-192x192.png
