+++ build-configs/malte/assets/favicon-48x48.png
