+++ build-configs/malte/assets/favicon-512x512.png
