+++ build-configs/malte/assets/favicon.ico
