+++ build-configs/malte/theme/colors.js
@@ -1,9 +1,9 @@
 // @flow
 
-import { commonColors, commonDarkColors } from '../common/colors'
-import type { ColorsType } from '../common/colors'
+import { commonColors, commonDarkColors } from '../../common/theme/colors'
+import type { ColorsType } from '../../common/theme/colors'
 
-const themeColor = '#fbda16'
+const themeColor = '#ff0000'
 
 export const darkColors: ColorsType = {
   themeColor,
