+++ src/modules/i18n/loadLocales.js
@@ -3,7 +3,7 @@
 import { reduce, forEach, merge } from 'lodash'
 import defaultLocales from '../../../locales/locales.json'
 import buildConfig from '../app/constants/buildConfig'
-import type { LocalesType } from '../../../build-configs/configs/BuildConfigType'
+import type { LocalesType } from '../../../build-configs/BuildConfigType'
 
 type TransformedLocalesType = { [language: string]: { [namespace: string]: { [key: string]: string } } }
