+++ src/setupTest.js
@@ -3,7 +3,7 @@
 import 'raf/polyfill'
 import { configure } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import integreatTestCmsBuildConfig from '../build-configs/configs/integreat-test-cms'
+import integreatTestCmsBuildConfig from '../build-configs/integreat-test-cms'
 
 configure({ adapter: new Adapter() })
